packages/excavator-mysql-migrate-cli/bin/get-global-options.ts
import { GlobalOptions } from './types'
import arg from 'arg'

export const getGlobalOptions = (argv_: string[]): GlobalOptions => {
  const argv = arg(
    {
      '--host': String,
      '--port': String,
      '--user': String,
      '--password': String,
      '--database': String,
    },
    {
      argv: argv_,
      permissive: true,
    },
  )

  const host = argv['--host'] || process.env['MYSQL_HOST'] || 'localhost'
  const port = parseInt(argv['--port'] || process.env['MYSQL_PORT'] || '3306', 10)
  const user = argv['--user'] || process.env['MYSQL_USER']
  const password = argv['--password'] || process.env['MYSQL_PASSWORD']
  const database = argv['--database'] || process.env['MYSQL_DATABASE']

  if (!user || !database) {
    throw new Error('user, database values are required')
  }

  return {
    host: host,
    port: port,
    user: user,
    password: password,
    database: database,
    extra: argv._,
  }
}
